debian/phaze-vpn/opt/phaze-vpn/extensibility-api/plugin-runtime.py
#!/usr/bin/env python3
"""
Plugin Runtime & Sandbox
Safe execution environment for user plugins
"""

import sys
import os
import json
from RestrictedPython import compile_restricted, safe_globals
from RestrictedPython.Guards import safe_builtins

class PluginRuntime:
    """Safe plugin execution environment"""
    
    def __init__(self):
        self.sandbox_globals = {
            '__builtins__': safe_builtins,
            '_print_': self.safe_print,
            '_getiter_': iter,
            '_getitem_': lambda obj, key: obj[key],
            'json': json,
            'datetime': __import__('datetime'),
        }
    
    def safe_print(self, *args, **kwargs):
        """Safe print function"""
        print(*args, **kwargs)
    
    def execute_plugin(self, code, context=None):
        """Execute plugin code in sandbox"""
        try:
            # Compile with restrictions
            byte_code = compile_restricted(code, '<plugin>', 'exec')
            
            if byte_code.errors:
                return {
                    'success': False,
                    'error': 'Compilation error',
                    'details': byte_code.errors
                }
            
            # Merge context
            exec_globals = self.sandbox_globals.copy()
            if context:
                exec_globals.update(context)
            
            # Execute
            exec(byte_code.code, exec_globals)
            
            # Get result
            result = exec_globals.get('result', {'message': 'Plugin executed'})
            
            return {
                'success': True,
                'result': result
            }
        except Exception as e:
            return {
                'success': False,
                'error': str(e)
            }

# Plugin API helpers
class PluginAPI:
    """API helpers for plugins"""
    
    @staticmethod
    def send_email(to, subject, body):
        """Send email via API"""
        import requests
        # Implementation
        pass
    
    @staticmethod
    def create_file(name, content):
        """Create file in storage"""
        import requests
        # Implementation
        pass
    
    @staticmethod
    def trigger_webhook(url, data):
        """Trigger webhook"""
        import requests
        # Implementation
        pass
