debian/phaze-vpn/opt/phaze-vpn/requirements.txt
# PhazeVPN GUI Client Requirements
# Python dependencies for the desktop GUI application

# Core dependencies
requests>=2.28.0
urllib3>=1.26.0

# System tray support (optional but recommended)
pystray>=0.19.0  # For system tray icon
Pillow>=9.0.0  # For image handling (required by pystray)

# Optional but recommended
bcrypt>=4.0.0  # For secure password hashing (has fallback if not available)

# Note: tkinter is usually included with Python, but may need:
# - Linux: python3-tk package
# - Windows: Usually included
# - macOS: Usually included

